feat: Remove item from available items in MarketplaceManager and hide it in ItemUI
The code changes in this commit modify the `ItemUI.cs` and `MarketplaceManager.cs` files. In `ItemUI.cs`, the `gameObject.SetActive(false)` line is added to hide the item UI after it is purchased. In `MarketplaceManager.cs`, the `AvailableItems` variable is changed from a 2D array to a jagged array, and the `RemoveItemFromAvailableItems` method is updated to remove the item from the `AvailableItems` array. This commit improves the functionality of the marketplace by removing purchased items from the available items list and hiding them in the UI.